requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
gunicorn==21.2.0
pymongo==4.5.0
numpy==1.26.0
pandas==2.1.1
scikit-learn==1.3.2
xgboost==2.0.3
python-multipart==0.0.6
python-dotenv==1.0.0
openpyxl==3.1.2